Locking moved to one file
